requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
httpx==0.27.2
python-multipart==0.0.17
python-dotenv==1.0.1

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.11.4
Pillow==11.0.0

# ML/AI
openai==1.58.1
numpy==2.2.0
tiktoken==0.8.0

# Vector Database
pinecone-client==5.0.1

# Graph Database
neo4j==5.26.0

# AWS
boto3==1.35.75
aioboto3==13.3.0

# Content Processing
beautifulsoup4==4.12.3
markdownify==0.14.1
langchain==0.3.13
langchain-openai==0.2.11
langchain-text-splitters==0.3.4

# Async Support
aiofiles==24.1.0
asyncio==3.4.3

# Logging & Monitoring
structlog==24.4.0
prometheus-fastapi-instrumentator==7.0.0

# Caching
aiocache==0.12.3
redis==5.2.1

# Utilities
python-jose[cryptography]==3.3.0
orjson==3.10.12
tenacity==9.0.0

# Development
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
httpx-mock==0.4.0
black==24.10.0
ruff==0.8.3
mypy==1.13.0
pre-commit==4.0.1